Moved weapon effect list into editor folder. Displaying list and form on the same page.
